refactor(pr07): relocate UoM constants to display, delete core/settings (G5f.3 + G5f.5)
Create src/settings/display/unitOfMeasure.mts with the live UoM symbols (imperialUnitOfMeasure, metricUnitOfMeasure, UNIT_OF_MEASURE_OPTIONS, WEIGHT_OPTIONS, plus type exports). Re-export from settings/display/index.mts. Drop the dead registerUnitOfMeasure function and the now-unreferenced UNIT_SETTINGS_KEY (the 'units' world setting is registered by display/registration.mts via DISPLAY_WORLD_KEYS.UNITS).

Update the sole importer (core/sheet/settingsStore.mts) to the new path. Delete src/settings/core/settings/ entirely.
